+++ packages/api-reference/src/types.ts
@@ -1,11 +1,21 @@
+import {
+  type OpenAPIV2,
+  type OpenAPIV3,
+  type OpenAPIV3_1,
+} from '@scalar/openapi-parser'
 import { type ThemeId } from '@scalar/themes'
 import type { MetaFlatInput } from '@unhead/schema'
 import type { HarRequest } from 'httpsnippet-lite'
-import { type OpenAPIV2, type OpenAPIV3, type OpenAPIV3_1 } from 'openapi-types'
-import { type DeepReadonly, type Slot } from 'vue'
+import { type Slot } from 'vue'
 
 export type ReferenceProps = {
   configuration?: ReferenceConfiguration
+}
+
+export type ReferenceLayoutProps = {
+  configuration: ReferenceConfiguration
+  parsedSpec: Spec
+  rawSpec: string
 }
 
 export type ReferenceConfiguration = {
@@ -75,21 +85,6 @@ export type SpecConfiguration = {
   url?: string
   /** Swagger/Open API spec */
   content?: string | Record<string, any> | (() => Record<string, any>)
-  /** The result of @scalar/swagger-parser */
-  preparsedContent?: Record<any, any>
-}
-
-/** Default reference configuration */
-export const DEFAULT_CONFIG: DeepReadonly<ReferenceConfiguration> = {
-  spec: {
-    content: undefined,
-    url: undefined,
-    preparsedContent: undefined,
-  },
-  proxy: undefined,
-  theme: 'default',
-  showSidebar: true,
-  isEditable: false,
 }
 
 export type GettingStartedExamples = 'Petstore' | 'CoinMarketCap'
@@ -279,6 +274,11 @@ export type Server = {
   variables?: ServerVariables
 }
 
+export type TagGroup = {
+  name: string
+  tags: string[]
+}
+
 export type SecurityScheme =
   | Record<string, never> // Empty objects
   | OpenAPIV2.SecuritySchemeObject
@@ -293,21 +293,23 @@ export type Webhooks = Record<
 >
 
 export type Spec = {
-  tags?: Tag[]
-  info:
+  'tags'?: Tag[]
+  'info':
     | Partial<OpenAPIV2.Document['info']>
     | Partial<OpenAPIV3.Document['info']>
     | Partial<OpenAPIV3_1.Document['info']>
-  host?: OpenAPIV2.Document['host']
-  basePath?: OpenAPIV2.Document['basePath']
-  schemes?: OpenAPIV2.Document['schemes']
-  externalDocs?: ExternalDocs
-  servers?: Server[]
-  components?: OpenAPIV3.ComponentsObject | OpenAPIV3_1.ComponentsObject
-  webhooks?: Webhooks
-  definitions?: Definitions
-  swagger?: OpenAPIV2.Document['swagger']
-  openapi?: OpenAPIV3.Document['openapi'] | OpenAPIV3_1.Document['openapi']
+  'host'?: OpenAPIV2.Document['host']
+  'basePath'?: OpenAPIV2.Document['basePath']
+  'schemes'?: OpenAPIV2.Document['schemes']
+  'externalDocs'?: ExternalDocs
+  'servers'?: Server[]
+  'components'?: OpenAPIV3.ComponentsObject | OpenAPIV3_1.ComponentsObject
+  'webhooks'?: Webhooks
+  'definitions'?: Definitions
+  'swagger'?: OpenAPIV2.Document['swagger']
+  'openapi'?: OpenAPIV3.Document['openapi'] | OpenAPIV3_1.Document['openapi']
+  'x-tagGroups'?: TagGroup[]
+  'security'?: OpenAPIV3.SecurityRequirementObject[]
 }
 
 export type AuthenticationState = {