+++ packages/swagger-editor/src/components/SwaggerEditor/SwaggerEditor.vue
@@ -1,13 +1,11 @@
 <script lang="ts" setup>
-import '@scalar/components/style.css'
-import { type ThemeId, ThemeStyles } from '@scalar/themes'
+import { ResetStyles, type ThemeId, ThemeStyles } from '@scalar/themes'
 import { computed, isRef, ref, watch } from 'vue'
 
 import {
   type OpenSwaggerEditorActions,
   type SwaggerEditorProps,
 } from '../../types'
-import ResetStyles from '../ResetStyles.vue'
 import SwaggerEditorHeader from './SwaggerEditorHeader.vue'
 import SwaggerEditorInput from './SwaggerEditorInput.vue'
 import SwaggerEditorNotification from './SwaggerEditorNotification.vue'
@@ -36,7 +34,7 @@ const formattedError = computed(() => {
   const error = isRef(props.error) ? props.error.value : props.error
 
   // Handle YAMLExceptions
-  if (error.startsWith('YAMLException:')) {
+  if (error.startsWith('YAMLParseError:')) {
     // Trim everything but the first line
     return error.split('\n')[0]
   }